django_app/requirements.txt
Django==4.1.2
django-environ==0.9.0
social-auth-app-django==5.0.0
pillow==9.2.0
gunicorn==20.1.0